+++ crates/client/src/workflow_handle.rs
@@ -6,7 +6,7 @@ use crate::{
     TerminateWorkflowInput, WorkflowCancelOptions, WorkflowDescribeOptions,
     WorkflowExecuteUpdateOptions, WorkflowExecutionStatus, WorkflowFetchHistoryOptions,
     WorkflowGetResultOptions, WorkflowQueryOptions, WorkflowSignalOptions,
-    WorkflowStartUpdateOptions, WorkflowTerminateOptions, WorkflowUpdateWaitStage,
+    WorkflowStartUpdateOptions, WorkflowTerminateOptions,
     errors::{
         WorkflowGetResultError, WorkflowInteractionError, WorkflowQueryError, WorkflowUpdateError,
     },
@@ -868,7 +868,6 @@ where
                 WorkflowStartUpdateOptions::builder()
                     .maybe_update_id(options.update_id)
                     .maybe_header(options.header)
-                    .wait_for_stage(WorkflowUpdateWaitStage::Completed)
                     .rpc_options(rpc_options.clone())
                     .build(),
             )
@@ -920,17 +919,6 @@ where
                             .codec()
                             .encode(&SerializationContextData::Workflow, unencoded_payloads?)
                             .await?;
-                        let lifecycle_stage = match options.wait_for_stage {
-                            WorkflowUpdateWaitStage::Admitted => {
-                                UpdateWorkflowExecutionLifecycleStage::Admitted
-                            }
-                            WorkflowUpdateWaitStage::Accepted => {
-                                UpdateWorkflowExecutionLifecycleStage::Accepted
-                            }
-                            WorkflowUpdateWaitStage::Completed => {
-                                UpdateWorkflowExecutionLifecycleStage::Completed
-                            }
-                        };
                         let update_id = options
                             .update_id
                             .unwrap_or_else(|| Uuid::new_v4().to_string());
@@ -941,7 +929,8 @@ where
                                 run_id,
                             }),
                             wait_policy: Some(WaitPolicy {
-                                lifecycle_stage: lifecycle_stage.into(),
+                                lifecycle_stage:
+                                    UpdateWorkflowExecutionLifecycleStage::Accepted.into(),
                             }),
                             request: Some(update::v1::Request {
                                 meta: Some(update::v1::Meta {